% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit1.R
\name{strsplit1}
\alias{strsplit1}
\title{TitleSplit a string}
\usage{
strsplit1(x, split)
}
\arguments{
\item{x}{A character with one element}

\item{split}{What to split on}
}
\value{
A character verctor
}
\description{
TitleSplit a string
}
\examples{
x <- "alpha,beta,cha"
strsplit1(x, ",")
}
